Spam filter for providers+level
